refactor(package-name): moved the name builder to the package directory
